sdl: run RunLoop's updateFunc on the calling goroutine on js; more bindings
RunLoop redesign (js): emscripten_set_main_loop invoked updateFunc through a
js.FuncOf event goroutine. When user code blocks there (input-wait loops that
sleep across frames), the Go wasm scheduler never pauses back to the browser
event loop and the tab's main thread wedges inside findRunnable (confirmed
with a CDP Debugger.pause stack). requestAnimationFrame now just signals a
channel; updateFunc runs on the goroutine that called RunLoop, where blocking
pauses cleanly. RAF pacing semantics for non-blocking (game-style) callbacks
are unchanged.

Also: un-stub HasEvent and SetRenderLogicalPresentation; platform-split
GetKeyboardState (SDL returns a pointer to its internal state array, which the
js path must copy out of the Emscripten heap — the shared PtrToSlice view only
works natively).

Verified: goxpyriment's parity_decision experiment runs end-to-end in headless
Chrome via CDP-dispatched key events — 10 trials, reaction times recorded,
CSV + info files downloaded through the browser.
