+++ CLAUDE.md
@@ -24,9 +24,16 @@ filesystem `replace github.com/Zyko0/go-sdl3 => ../go-sdl3-wasm` is seamless.
   `go run ./cmd/wasmsdl serve ./path/to/app`.
 - Go bindings for js live in `*/*_functions_js.go` as `iXxx` closures that
   marshal args and call `js.Global().Get("Module").Call("_SDL_Xxx", ...)`.
-- The program's frame loop must go through `sdl.RunLoop`; on js this drives the
-  frame with `emscripten_set_main_loop` (a hand-written `for { PollEvent; Delay }`
-  loop freezes the browser tab).
+- The program's frame loop must go through `sdl.RunLoop`. On js it paces
+  iterations with `requestAnimationFrame` but runs the callback **on the
+  goroutine that called RunLoop** (the main goroutine), NOT via
+  `emscripten_set_main_loop`. Rule learned the hard way: user code that
+  *blocks* (sleep/channel waits) inside a `js.FuncOf` callback pins the tab's
+  main thread inside the Go scheduler (`findRunnable` never pauses back to the
+  event loop), while the same blocking code on the main goroutine pauses
+  cleanly. So JS callbacks must only signal channels; all user code runs on
+  Go-owned goroutines. (Also: a hand-written `for { PollEvent; SDL_Delay }`
+  loop freezes the tab — `SDL_Delay` is a no-op on js; use `time.Sleep`.)
 - The js bindings are generated by `cmd/jsgen`, which emits
   `panic("not implemented on js")` as the first line of every function. A
   function only works once that panic is removed and its argument/return
